fix(sd): stop two #924 comments overstating what was measured
Pre-PR self-review of the branch diff. Both findings are the same class and
it is the one this repo's debugging-discipline rule is written against: a
comment presenting an ESTIMATE, or a test that was not run, as a measurement.

- SDCardWrite()'s cost note said the A/B soak "compares SdDroppedBytes at the
  enforced cap against a no-manifest control of the same shape". No such
  control exists: one firmware image cannot be its own control, and
  test_924_sd_manifest.py measures (rotating - not-rotating) / rotations on a
  single image against test_824's 512 B/rotation gate instead. The comment now
  says that, cites the measured pre-#924 reference (0.0-2.1 B/rotation over two
  trials on NQ1 7E2898F46200E8A7), and states plainly that the paragraph bounds
  what the manifest adds to the ROTATION window rather than measuring the CRC's
  steady-path cost.
- The manifest block's no-per-line-FileSync note repeated the same phrase. It
  now names what the close actually guarantees -- SYS_FS_FileClose makes the
  manifest durable, so a cleanly stopped session has no unflushed window -- and
  narrows the stated consequence to the abnormal-stop case it really is.

Also from the same review: two SYS_FS_Error() log arguments cast to (int),
matching the stricter sites already in this file rather than passing a bare
enum to %d; and tests/host/.gitignore gains run_924_tests, which it otherwise
misses exactly as it missed run_913_tests.

Re-verified after the edits: full XC32 build clean (-Werror -Wall, 0 errors,
hex regenerated in this worktree), cppcheck 0 findings with the baseline
unchanged, tests/host 15/15 suites 0 failed.

Refs #924

+++ firmware/src/services/sd_card_services/sd_card_manager.c
@@ -518,10 +518,19 @@ static int SDCardWrite() {
      * lookups per byte), and writeBuffer is the COHERENT (KSEG1, uncached)
      * pool allocation, so each byte load is an uncached read with no cache-
      * line benefit. At the ~500 KB/s SD ceiling that is single-digit percent
-     * of one core on a task that spends its time waiting on SPI. The A/B soak
-     * in #924's acceptance criteria is what tests that claim rather than
-     * asserting it -- it compares SdDroppedBytes at the enforced cap against a
-     * no-manifest control of the same shape. */
+     * of one core on a task that spends its time waiting on SPI.
+     *
+     * THAT IS AN ESTIMATE, AND THE TEST THAT CHECKS IT IS AN ON-DEVICE A/B,
+     * NOT A COMPARISON AGAINST A NO-MANIFEST IMAGE. #924's acceptance wording
+     * asks for the latter; one firmware image cannot be its own control, so
+     * test_924_sd_manifest.py measures (SdDroppedBytes rotating -
+     * SdDroppedBytes not-rotating) / rotations on one image and compares it
+     * against test_824's 512 B/rotation gate on test_824's shape. The
+     * pre-#924 reference for that figure, measured on NQ1 7E2898F46200E8A7
+     * over two trials, is 0.0-2.1 B/rotation. Do not read this paragraph as a
+     * measurement of the CRC's cost: it bounds what the manifest adds to the
+     * ROTATION window, and the steady-path cost shows up instead as the
+     * arms' streamed-at-rate check. */
     if (writeLen > 0) {
         size_t folded = (size_t)writeLen;
         if (folded > gSDCardData.writeBufferLength) {
@@ -1513,8 +1522,10 @@ static void sd_AbandonRotationWindow(const char* why) {
  * window, which #757/#822/#824 spent three issues keeping short because the
  * encoder is filling the ring throughout it; a metadata flush per rotation
  * would add SD writes to exactly the window whose cost #924's own acceptance
- * criteria measure (no new SdDroppedBytes versus a no-manifest control). The
- * close at session end flushes everything. The stated consequence: a session
+ * criteria measure. The close at session end flushes everything, and
+ * SYS_FS_FileClose is what makes the manifest durable -- there is no window in
+ * which a cleanly-stopped session leaves it unflushed. The stated consequence
+ * of having no per-line sync is narrower than that: a session
  * ended by power loss or a card yank can leave a manifest short of its last
  * line(s) -- the stream files it does name are still fully described, and a
  * host that wants a CRC for an unnamed file can still ask
@@ -1546,7 +1557,7 @@ static void sd_OpenSessionManifest(void) {
                                                  SYS_FS_FILE_OPEN_WRITE);
     if (gSDCardData.manifestHandle == SYS_FS_HANDLE_INVALID) {
         LOG_E("[SD] #924 could not open manifest '%s' (err=%d) - session "
-              "continues without one", path, SYS_FS_Error());
+              "continues without one", path, (int)SYS_FS_Error());
         return;
     }
     gSDCardData.manifestLines = 0;
@@ -1598,7 +1609,8 @@ static void sd_CloseSessionManifest(void) {
         return;
     }
     if (SYS_FS_FileClose(gSDCardData.manifestHandle) == SYS_FS_RES_FAILURE) {
-        LOG_E("[SD] #924 failed to close manifest (err=%d)", SYS_FS_Error());
+        LOG_E("[SD] #924 failed to close manifest (err=%d)",
+              (int)SYS_FS_Error());
     } else {
         LOG_I("[SD] #924 manifest closed: %u file(s) recorded",
               (unsigned)gSDCardData.manifestLines);
